[REFACTOR] Extract api integration code out of the repository class
In the new use case (being able to cache info) we could obtain info from a cache or
from the API and that should be the responsibility of the repository class. API
integration code has been moved to its own class (that is tested using integration
tests) and repository right now is just glue code passing a messages (it does not
need tests).